Add `playlists` endpoint
- bump `yayti` to `v0.3.0`
- switch to using `mimalloc` because it more readily deallocates
- update README
- add `playlists-path` setting for local playlists
- add tiny frontend for viewing if the API is working (correspond to invidious frontend links `/watch` and `/playlist`)
- add icon to project so that executable can be easier to spot in task manager
- add branch & commit info to `stats` endpoint

+++ src/helpers.rs
@@ -79,6 +79,7 @@ impl DbWrapper {
       preference: DbType::MongoDb
     }
   }
+  // TODO ✏ add better error handling for timeouts per https://docs.rs/mongodb/latest/mongodb/#warning-about-timeouts--cancellation
   pub async fn seek_for_json(&self, collection_name: &str, key: &str) -> Option<Value> {
     match self.preference {
       DbType::UnQLite => {
@@ -141,7 +142,7 @@ pub async fn get_previous_data(collection: &str, key: &str, db: &DbWrapper, app_
           Some(timestamp) => {
             let current_timestamp = Utc::now().timestamp();
             let offset = current_timestamp - timestamp;
-            if offset as i32 > app_settings.cache_timeout {
+            if offset as u64 > app_settings.cache_timeout {
               db.delete(collection, key).await;
               None
             } else {
